Time: 102 ms (56.02%), Space: 13.2 MB (78.12%) - LeetHub

+++ 18-4sum/18-4sum.cpp
@@ -10,7 +10,7 @@ public:
         //very imp to remove duplicates
         if(i>0 && nums[i]==nums[i-1])
             continue;
-        int target_3 = target-nums[i];
+        long long target_3 = (long long)target-nums[i];
         for(int j=i+1;j<n;j++)
         {
             //very imp to remove duplicates
